scraper/requirements.txt
praw
pandas
spacy
fuzzywuzzy
mysql-connector-python
yfinance